cpp: Don't send empty compilationDatabasePath on initialization
If there is no active build configuration, we currently send an empty
string as the compilationDatabasePath to clangd in the
initializationOptions.  The intention was that clangd would interpret it
as a way to deselect any previously specified compilationDatabasePath
(in other words, go back to the initial state of not explicitly defined
compile commands path).  However, as of today, clangd understands an
empty string as a regular path, and it doesn't give good results.  The
compile_commands.json file is not found is clangd's current working
directory has changed (which happens when it analyzes a file in a
different directory).

Clearly, a fix is needed in clangd: we need a way to specify "no
compilation database path".  But in the mean time, we can mitigate this
problem by not sending anything during the initialization phase if there
is not active build configuration.  Clangd will keep its initial state
of "no compilation database path".

The case of choosing a build config and then choosing no build config
will still be broken though, as we will still send an empty string.
This will require the fix in clangd mentioned previously.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>