GH-2882: Replaced the `range` with the `selectionRange` in the Outline.
So that when one selects a tree node in the Outline view
only the name of the symbol will be selected.
Previously, it was the entire enclosing range of the symbol.
For instance, the method/function body, instead of the name.

Fixed a bug in the Outline view widget, when reconciling the tree nodes.

Added support to update the Outline view selection based on the
selection from the currently active editor.

Closes: #2882.
Closes: #2883.

Signed-off-by: Akos Kitta <kittaakos@typefox.io>