Move tasks test-resources to top-level
I tried to run the task tests by running a watch on the task package,
then running the compiled javascript spec file directly, with:

    ./node_modules/.bin/mocha --opts configs/mocha.opts packages/task/lib/node/task-server.slow-spec.js

This doesn't work, because the test resources are in
packages/task/src/node/test-resources, while the above command looks for
them in packages/task/lib/node/test-resources.  The alternative is to
run mocha on the src/node/task-server.slow-spec.ts file, but it takes a bit
longer, which is why I prefer to have a watch and execute the js file.

This patch makes it work for both methods.

Signed-off-by: Simon Marchi <simon.marchi@ericsson.com>