[WIP] use dash-licenses tool from Eclipse Foundation
to validate our repo's runtime dependencies for license compatibility,
according to the Eclipse Foundation's rules.

In this PR we add the possibility to run the "dasn-licenses" tool, that's
being developed by the Eclipse Foundation. It will eventually replace our
manual production dependencies license check. As far as I can tell,
dash-licenses has not been released yet, so as a start, we can add its
repo as a git sub-mosule and built it in place, before using it.

Dash-licenses is added as a git submodule. It's mostly intended for CI.
Eventually we'll be able to install the tool from a package (maven central,
npm, ...) :

$  yarn prepare:licenseCheck

Run the license check tool. This assumes the jar be found with this glob:
dash-licenses/target/org.eclipse.dash.licenses* :

$ yarn licenseCheck

The strange JVM srgument is to avoid a failure of "dash-licenses" at runtime
with this problem:
https://github.com/structurizr/java/issues/119

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>